# stewardess_openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stewardess MCP Service",
    "description": "Local source-code repository tools for AI agents. Provides file reading, search, editing, git, and command execution.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://office:55702"
    }
  ],
  "paths": {
...
    "/api/edit/write": {
      "post": {
        "operationId": "Edit_WriteFile",
        "summary": "Overwrites (or creates) a file with the supplied content.",
        "description": "Overwrites (or creates) a file with the supplied content.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/create-file": {
      "post": {
        "operationId": "Edit_CreateFile",
        "summary": "Creates a new file; optionally fails if the file already exists.",
        "description": "Creates a new file; optionally fails if the file already exists.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/create-directory": {
      "post": {
        "operationId": "Edit_CreateDirectory",
        "summary": "Creates a directory (with optional parent creation).",
        "description": "Creates a directory (with optional parent creation).",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDirectoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/delete-file": {
      "post": {
        "operationId": "Edit_DeleteFile",
        "summary": "Deletes a single file. A backup is created (and a rollback token returned)\r\nunless options.createBackup is false.\r\nWhen ApprovalRequiredForDestructive is enabled an approval token is required.",
        "description": "Deletes a single file. A backup is created (and a rollback token returned)\r\nunless options.createBackup is false.\r\nWhen ApprovalRequiredForDestructive is enabled an approval token is required.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/delete-directory": {
      "post": {
        "operationId": "Edit_DeleteDirectory",
        "summary": "Deletes a directory. Requires recursive=true for non-empty directories.",
        "description": "Deletes a directory. Requires recursive=true for non-empty directories.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDirectoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/append": {
      "post": {
        "operationId": "Edit_AppendFile",
        "summary": "Appends content to the end of an existing file.",
        "description": "Appends content to the end of an existing file.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/replace-text": {
      "post": {
        "operationId": "Edit_ReplaceText",
        "summary": "Replaces occurrences of a literal string within a file.",
        "description": "Replaces occurrences of a literal string within a file.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/replace-lines": {
      "post": {
        "operationId": "Edit_ReplaceLines",
        "summary": "Replaces a range of lines in a file with new content.",
        "description": "Replaces a range of lines in a file with new content.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceLinesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/edit/patch": {
      "post": {
        "operationId": "Edit_PatchFile",
        "summary": "Applies a unified diff patch to a single file.",
        "description": "Applies a unified diff patch to a single file.",
        "tags": [
          "Edit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WriteFileRequest": {
        "description": "Overwrites (or creates) a file with new content.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "content": {
            "description": "New full content of the file.",
            "type": "string"
          },
          "encoding": {
            "description": "Encoding to use when writing.  Supported values: \"utf-8\", \"utf-8-bom\",\r\n\"utf-16\", \"ascii\".  Defaults to \"utf-8\".",
            "type": "string"
          },
          "lineEnding": {
            "description": "Line ending style: \"lf\", \"crlf\", \"cr\", \"auto\" (preserve existing, default).",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "EditOptions": {
        "description": "Common options applied to every mutating operation.",
        "type": "object",
        "properties": {
          "dryRun": {
            "description": "When true, the operation is simulated and no files are changed.\r\nThe response contains the full proposed diff.",
            "type": "boolean"
          },
          "createBackup": {
            "description": "When true (default), a backup copy is created before overwriting.\r\nThe backup path is returned in {StewardessMCPService.Models.EditResult.BackupPath}.",
            "type": "boolean"
          },
          "changeReason": {
            "description": "Optional agent-supplied reason for the change, recorded in the audit log.",
            "type": "string"
          },
          "sessionId": {
            "description": "Agent-level session/correlation identifier for grouping related operations.",
            "type": "string"
          }
        }
      },
      "CreateFileRequest": {
        "description": "Creates a new file. Fails if the file already exists unless Overwrite is true.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "content": {
            "description": "Initial content of the new file.",
            "type": "string"
          },
          "encoding": {
            "description": "Encoding to use when writing the file (e.g. \"utf-8\").",
            "type": "string"
          },
          "overwrite": {
            "description": "When true, overwrites an existing file.",
            "type": "boolean"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "CreateDirectoryRequest": {
        "description": "Creates a new directory at the given path.",
        "type": "object",
        "properties": {
          "path": {
            "description": "Directory path relative to repository root.",
            "type": "string"
          },
          "createParents": {
            "description": "When true, creates all missing parent directories.",
            "type": "boolean"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
...
      "DeleteFileRequest": {
        "description": "Deletes a single file within the repository.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "approvalToken": {
            "description": "Confirmation token returned by a prior preview_changes call.\r\nRequired when ApprovalRequiredForDestructive is enabled.",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "DeleteDirectoryRequest": {
        "description": "Deletes a directory, optionally recursively.",
        "type": "object",
        "properties": {
          "path": {
            "description": "Directory path relative to repository root.",
            "type": "string"
          },
          "recursive": {
            "description": "When false (default), deletion fails if the directory is non-empty.",
            "type": "boolean"
          },
          "approvalToken": {
            "description": "Approval token required when ApprovalRequiredForDestructive is enabled.",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "AppendFileRequest": {
        "description": "Appends content to the end of an existing file.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "content": {
            "description": "Text to append.",
            "type": "string"
          },
          "ensureNewLine": {
            "description": "When true, ensures the content starts on a new line.",
            "type": "boolean"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "ReplaceTextRequest": {
        "description": "Replaces all occurrences of a literal string within a file.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "oldText": {
            "description": "The literal text to find.",
            "type": "string"
          },
          "newText": {
            "description": "The replacement text.",
            "type": "string"
          },
          "ignoreCase": {
            "description": "When true, performs a case-insensitive search.",
            "type": "boolean"
          },
          "maxReplacements": {
            "format": "int32",
            "description": "When &gt; 0, limits the number of replacements made.",
            "type": "integer"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "ReplaceLinesRequest": {
        "description": "Replaces a range of lines in a file with new content.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "startLine": {
            "format": "int32",
            "description": "1-based start line (inclusive).",
            "type": "integer"
          },
          "endLine": {
            "format": "int32",
            "description": "1-based end line (inclusive).",
            "type": "integer"
          },
          "newContent": {
            "description": "Replacement text (can span multiple lines).",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
      "PatchFileRequest": {
        "description": "Applies a unified diff patch to a single file.",
        "type": "object",
        "properties": {
          "path": {
            "description": "File path relative to repository root.",
            "type": "string"
          },
          "patch": {
            "description": "Unified diff text (--- a/ +++ b/ format).",
            "type": "string"
          },
          "fuzzFactor": {
            "format": "int32",
            "description": "Number of context lines for fuzzy matching (default 3).",
            "type": "integer"
          },
          "options": {
            "$ref": "#/components/schemas/EditOptions",
            "description": "Write options such as dry-run mode and backup control."
          }
        }
      },
...
    }
  }
}